requirements.txt
asgiref==3.9.0
attrs==25.3.0
beautifulsoup4==4.12.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework==3.14.0
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
pycparser==2.22
PySocks==1.7.1
python-decouple==3.8
pytz==2025.2
requests==2.31.0
selenium==4.15.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
tzdata==2025.2
urllib3==2.5.0
wsproto==1.2.0